langchain_12_25_2025/requirements.txt
ipykernel
python-dotenv
langchain
langchain-community
langchain_community 
pypdf
bs4
arxiv
pymupdf>=1.23
wikipedia
langchain-text-splitters
langchain-openai
chromadb
langchain_ollama
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette